Fix compile error on windows.
R=yangguo@chromium.org

BUG=
TEST=

Review URL: https://chromiumcodereview.appspot.com/9424032

git-svn-id: http://v8.googlecode.com/svn/branches/bleeding_edge@10752 ce2b1a6d-e550-0410-aec6-3dcde31c8c00

+++ src/objects.h
@@ -1012,7 +1012,8 @@ class Smi: public Object {
   void SmiVerify();
 #endif
 
-  static const int kMinValue = (-1U << (kSmiValueSize - 1));
+  static const int kMinValue = 
+      (static_cast<unsigned int>(-1)) << (kSmiValueSize - 1);
   static const int kMaxValue = -(kMinValue + 1);
 
  private:
